requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Google AI + ADK
google-adk==1.28.0
google-cloud-aiplatform==1.71.1
google-genai==1.9.0

# MCP
fastmcp==2.3.3

# Database
psycopg2-binary==2.9.9

# Validation + config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2

# Retry logic
tenacity==8.5.0